util: move WaitGroupWithError to util from utilccl
Closes #13117.
